Add files via upload
This Real Estate Price Prediction App uses machine learning algorithms to predict the price of a property based on key features such as the number of bedrooms, bathrooms, and the size of the property in square feet. The app aims to help users get an estimated price for their property, offering valuable insights for potential buyers and sellers in the real estate market.